feat: use fogo.io/api/token-metadata instead of metaplex (#85)
Reading the metaplex accounts is expensive and unnecessary for loading metadata.
This PR instead moves to using the new token-metadata endpoint on the fogo
website.

Note we still use metaplex when building the intent message, because we have to
use metaplex on chain when reading it and it's important the intent message is
built and read using the same metadata source.

This PR also ensures that tokens in the token list have a stable sort order, and
that tokens that have no metadata are sorted after those that do have metadata